requirements.txt
Django==2.0.13
django-annoying==0.10.3
django-debug-toolbar==1.11
django-crispy-forms==1.7.2
django-downloadview==1.9
django-js-reverse==0.9.1
mysqlclient==1.3.9
Pillow==4.0.0
sorl-thumbnail==12.4.1
